crawler/pyproject.toml
[tool.poetry]
name = "url-to-llm-crawler"
version = "0.1.0"
description = "Playwright-powered incremental web crawler"
authors = ["URL to LLM Team"]

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.40.0"
httpx = "^0.25.2"
asyncpg = "^0.29.0"
aioboto3 = "^12.0.0"
redis = "^5.0.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
robotexclusionrulesparser = "^1.7.1"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
black = "^23.12.1"
ruff = "^0.1.9"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
env = [
    "DATABASE_URL=postgresql://postgres:postgres@localhost:5432/test_db",
    "REDIS_URL=redis://localhost:6379",
    "S3_ENDPOINT=http://localhost:9000",
    "S3_ACCESS_KEY=minioadmin",
    "S3_SECRET_KEY=minioadmin",
    "S3_BUCKET=test-bucket",
    "PLAYWRIGHT_CHROMIUM=true",
    "ENVIRONMENT=test"
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]